win32: really fix the DllMain signature
After 9ca790a6414b43c81ed0d539e9d200cc25ea4b0e the DWORD turned into
an unsigned, but it should be an unsigned long.

+++ modules/access/wasapi.c
@@ -75,7 +75,7 @@ static msftime_t GetQPC_100ns(void)
 
 static msftime_t (*get_qpc)(void);
 
-int __stdcall DllMain(void *dll, unsigned reason, void *reserved)
+int __stdcall DllMain(void *dll, unsigned long reason, void *reserved)
 {
     (void) dll;
     (void) reserved;
